test: skip autofix acceptance tests no longer exposed
Skip acceptance tests for autofix feature since it is no longer
exposed as a user-facing feature. This prevents test failures
and reflects the current state of the application while keeping
the test code for potential future re-enablement.
